requirements.txt
# Web stack
fastapi==0.104.1
starlette==0.27.0
httpx==0.25.1

# Core libs
pydantic>=2.7,<3
pydantic-settings>=2.2,<3
tenacity==8.2.3
python-dotenv==1.0.0
requests==2.31.0
uvicorn==0.24.0
pysnow==0.7.17

# Testing dependencies
pytest==8.4.2
pytest-asyncio>=0.23
pytest-cov==4.1.0
